pyproject.toml
[tool.poetry]
name = "qserverctrl"
version = "0.1.0"
description = ""
authors = ["Yi Cao <me@ycao.top>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tencentcloud-sdk-python-cvm = "^3.0.925"
tencentcloud-sdk-python-common = "^3.0.925"
mcstatus = "^10.0.3"
websockets = "^11.0.3"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
